requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
supabase>=2.10.0
pydantic>=2.10.4
pydantic-settings>=2.0.0
openai>=1.58.1
pytest>=7.0.0
httpx>=0.24.0 
streamlit==1.41.1